Stop a compiler warning

+++ src/qtrait.cpp
@@ -95,7 +95,7 @@ void pyQuanTrait::qtrait(individual * ind, ULONG gen, vectorf & traits)
 
 	DBG_ASSERT(args, RuntimeError, "Failed to create a parameter tuple");
 
-	for (size_t i = 0; i < m_func.numArgs(); ++i) {
+	for (int i = 0; i < m_func.numArgs(); ++i) {
 		const string & arg = m_func.arg(i);
 		if (arg == "ind")
 			PyTuple_SET_ITEM(args, i, pyIndObj(static_cast<void *>(ind)));
